provider: Handle relative paths in the uri
This allows to properly handle absolute and relative paths. The paths
in a QGIS project file are expected to be relative by
default (relative to the QGIS project file path). In order to properly
handle it, it is necessary to:
- convert paths to relative paths when the project is saved
- convert paths to absolute paths when the project is read

`DuckdbProviderMetadata.decodeUri` is called when the uri from a QGIS
project file is read. If the path of the uri has been stored as
relative, it needs to be converted into an absolute path to be
properly handled by the provider.
This is achieved by `QgsProject.instance().pathResolver().readPath`
which converts the path into an absolute one.

When a project is saved, the path of the uri needs to be converted
into a relative path. The save logic automatically calls
`DuckdbProviderMetadata.absoluteToRelativeUri`. It convert the uri
into a relative one:
1. the uri is decoded
2. the path is converted to a relative one by calling
`pathResolver().writePath`
3. a new uri is encoded with the relative path
However, this mechanism does not work before QGIS 3.30 because
`absoluteToRelativeUri` did not exist. Prior to this version, it is
not possible to save a project with relative urls from a python
provider.