provider_metadata: Do not change path in decodeuri for QGIS >= 3.30
This is a follow up of e51fff344a2a0a83e58108d4f653381e6b05fa18

The readPath call was introduced to avoid a limitation of older QGIS
versions (QGIS < 3.30). This call ensures to have an absolute path
whether 'path' is absolute or not. However, there is no need to call
it for recent QGIS versions.

The logic to parse an uri and convert the path from relative to
absolute is:
1. call `QGsVectorLayer::decodedSource()` to parse the uri and
convert the path with `readPath`
2. call `QgsProviderMetadata.decodeUri()` to parse the final uri which
already contains an absolute one.

However, prior to QGIS 3.30, this does not work for duckdb
provider. Indeed, the behavior of each provider was hardcoded in the
function `QGsVectorLayer::decodedSource()` and it could not handle
duckdb provider.
Since, QGIS 3.30, this has been delegated to
QgsProviderMetadata::relativeToAbsoluteUri. This allows each provider
to have its own behavior and fix the issue for duckdb provider.

Since it is not possible to override QGsVectorLayer::decodedSource(),
prior to QGIS 3.30, the uri used to call `decodeUri` contains a
relative path instead of an absolute one. By calling `readPath`, this
solves the issue.