duckdb_feature_iterator: Add support for SubsetOfAttributes flag
When the `SubsetOfAttributes` flag is set on a request, it means that
only a subset of the attributes need to be retrieved. The list of
attributes is stored as indexes by `request.subsetOfAttributes()`.

In that case, two changes need to be applied to the iterator:
- update the initial duckdb `select` request to only retrieve the
requested attributes
- in the `fetchFeature`, only call `feature.setAttribute()` on the
necessary attributes. If all the attributes have been retrived
`feature.setAttributes()` can directly be called.

If no attribute is requested, this can speed up the iterator
significantly because the setAttribute/setAttributes calls can take a
lot of time.