test_plg_metadata: Fix min,max version comparison
A minimum or maximum version as defined in the metadata.txt file is
not necessarily a float, ie. '3.34'. It can be a semver version such
as '3.34.5'.
In the latter case, the original string cannot be converted to a
float and the test fails.

This issue is fixed by using packaging.version.parse. It allows to
correctly parse the min/max version string and compare them.