duckdb_feature_iterator: Correctly handle date/time fields
QGIS cannot properly handle attributes coming from python `datetime`
objects because the generated `QVariant` cannot be parsed.
In that case, it is necessary to convert them to their Qt counterpart
which are correctly handled by `QVariant`.

This is achieved by detecting which fields are date or time. If such
fields exist, a converter is associated. It is used to convert an
attribute to its Qt counterpart when the feature are fetched.

Co-authored-by: Jean Felder <jean.felder@oslandia.com>

Closes: https://gitlab.com/Oslandia/qgis/qduckdb/-/issues/23