refactor: Switch to Qgis.MessageLevel enum for log messsage
`log_level` parameter expects a `Qgis.MessageLevel` value. This works
in both qt5 and qt6. In Qt5, this also works when an integer is
provided because it is able to convert it in the corresponding
`Qgis.MessageLevel` value. However, this does not work in Qt6 anymore.

PlgLogger.log has already been updated to convert the integer to a
`Qgis.MessageLevel` enum which works in Qt5 and Qt6.
However, one should switch to Qgis.MessageLevel everywhere because the
enum value is easier to understand than a magic integer value.