fix(tests): Isolate qgis module import between tests
The unit tests using a spatial filter (`setFilterRect`) do not work
when `qgis.testing.start_app` is called. This probably a conflict
between one of the static library used by duckdb and the dynamic
system libraries imported by QGIS modules.

Avoid this issue by unloading the QGIS modules at the end of a test
module. This way, the offending modules are not loading when the unit
tests needing `setFilterRect` are called.